+++ examples/ansibleplaybook-signals-and-cancellation/ansibleplaybook-signals-and-cancellation.go
@@ -7,9 +7,9 @@ import (
 	"os/signal"
 
 	"github.com/apenella/go-ansible/pkg/execute"
+	"github.com/apenella/go-ansible/pkg/execute/result/transformer"
 	"github.com/apenella/go-ansible/pkg/options"
 	"github.com/apenella/go-ansible/pkg/playbook"
-	"github.com/apenella/go-ansible/pkg/stdoutcallback/results"
 )
 
 func main() {
@@ -32,7 +32,7 @@ func main() {
 		Options:           ansiblePlaybookOptions,
 		Exec: execute.NewDefaultExecute(
 			execute.WithTransformers(
-				results.Prepend("Go-ansible example"),
+				transformer.Prepend("Go-ansible example"),
 			),
 		),
 	}